order cart 2

+++ db/migrate/20180508074514_create_promotions.rb
@@ -2,7 +2,7 @@ class CreatePromotions < ActiveRecord::Migration[5.1]
   def change
     create_table :promotions do |t|
       t.string :name
-      t.float :percent
+      t.decimal :percent
       t.string :description
       t.datetime :start_date
       t.datetime :end_date
