+++ db/migrate/20180508074832_create_order_items.rb
@@ -4,7 +4,7 @@ class CreateOrderItems < ActiveRecord::Migration[5.1]
       t.integer :order_id
       t.integer :product_id
       t.integer :quantity
-      t.float :total
+      t.decimal :total
 
       t.timestamps
     end
